ssh: NewKeys is created with CIPHER_NONE, use in TransportLayerIO; use ints
CipherSpec.CIPHER_NONE is a none-cipher with block size 8.
It's used in TransportLayerIO to elimiate the need of use of kex().

Also, the int is more than enough to store block size, etc.

+++ src/main/java/me/laszloattilatoth/jada/proxy/ssh/kex/algorithm/CipherRegistry.java
@@ -40,7 +40,7 @@ public class CipherRegistry {
         return byId(nameWithId.nameId());
     }
 
-    private static void put(String name, long blockSize, long keyLen, long ivLen, long authLen, long flags) {
+    private static void put(String name, int blockSize, int keyLen, int ivLen, int authLen, long flags) {
         int nameId = Name.getNameId(name);
         list.put(nameId, new CipherSpec(name, nameId, blockSize, keyLen, ivLen != 0 ? ivLen : blockSize, authLen, flags));
     }
